requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.7.0
djangorestframework==3.16.1
pathlib2==2.3.7.post1
psycopg2-binary==2.9.11
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
